+++ vscode-ui/src/config.ts
@@ -64,6 +64,7 @@ export interface CommandExecutionContext {
   data_dirs?: string[];
   cache_root?: string;
   init?: (string | [string, Record<string, any>])[];
+  auto_reload?: boolean;
   result_refresh_period?: number | null;
   status_refresh_period?: number | null;
 }